+++ pmd-core/src/test/java/net/sourceforge/pmd/cli/CoreCliTest.java
@@ -7,13 +7,16 @@ package net.sourceforge.pmd.cli;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.containsStringIgnoringCase;
+import static org.hamcrest.Matchers.endsWith;
 import static org.hamcrest.Matchers.not;
+import static org.hamcrest.Matchers.startsWith;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.FilterOutputStream;
 import java.io.IOException;
@@ -23,6 +26,8 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.List;
 import java.util.logging.Logger;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
@@ -39,7 +44,10 @@ import org.junit.rules.TemporaryFolder;
 
 import net.sourceforge.pmd.PMD;
 import net.sourceforge.pmd.PMD.StatusCode;
+import net.sourceforge.pmd.RuleContext;
 import net.sourceforge.pmd.junit.JavaUtilLoggingRule;
+import net.sourceforge.pmd.lang.ast.Node;
+import net.sourceforge.pmd.lang.rule.MockRule;
 import net.sourceforge.pmd.util.IOUtil;
 
 /**
@@ -48,6 +56,7 @@ import net.sourceforge.pmd.util.IOUtil;
 public class CoreCliTest {
 
     private static final String DUMMY_RULESET = "net/sourceforge/pmd/cli/FakeRuleset.xml";
+    private static final String DUMMY_RULESET_WITH_VIOLATIONS = "net/sourceforge/pmd/cli/FakeRuleset2.xml";
     private static final String STRING_TO_REPLACE = "__should_be_replaced__";
 
     @Rule
@@ -65,12 +74,8 @@ public class CoreCliTest {
 
     @Before
     public void setup() throws IOException {
-        // set current directory to wd
-        Path root = tempRoot();
-        System.setProperty("user.dir", root.toString());
-
         // create a few files
-        srcDir = Files.createDirectories(root.resolve("src"));
+        srcDir = Files.createDirectories(tempRoot().resolve("src"));
         writeString(srcDir.resolve("someSource.dummy"), "dummy text");
         // reset logger?
         Logger.getLogger("net.sourceforge.pmd");
@@ -128,6 +133,133 @@ public class CoreCliTest {
         runPmdSuccessfully("--no-cache", "--dir", srcDir, "--rulesets", DUMMY_RULESET, "--report-file", reportFile);
 
         assertTrue("Report file should have been created", Files.exists(reportFile));
+    }
+
+    @Test
+    public void testNoRelativizeWithAbsoluteSrcDir() {
+        assertTrue("srcDir should be absolute", srcDir.isAbsolute());
+        startCapturingErrAndOut();
+        runPmd(StatusCode.VIOLATIONS_FOUND, "--no-cache", "--dir", srcDir, "--rulesets", DUMMY_RULESET_WITH_VIOLATIONS);
+
+        assertThat(outStreamCaptor.getLog(), containsString(srcDir.resolve("someSource.dummy").toString()));
+    }
+
+    @Test
+    public void testNoRelativizeWithRelativeSrcDir() throws IOException {
+        // Note, that we can't reliably change the current working directory for the current java process
+        // therefore we use the current directory and make sure, we are at the correct place - in pmd-core
+        Path cwd = Paths.get(".").toRealPath();
+        assertThat(cwd.toString(), endsWith("pmd-core"));
+        String relativeSrcDir = "src/test/resources/net/sourceforge/pmd/cli/src";
+        assertTrue(Files.isDirectory(cwd.resolve(relativeSrcDir)));
+
+        startCapturingErrAndOut();
+        runPmd(StatusCode.VIOLATIONS_FOUND, "--no-cache", "--dir", relativeSrcDir, "--rulesets", DUMMY_RULESET_WITH_VIOLATIONS);
+
+        assertThat(outStreamCaptor.getLog(), containsString("\n" + IOUtil.normalizePath(relativeSrcDir + "/somefile.dummy")));
+    }
+
+    @Test
+    public void testNoRelativizeWithRelativeSrcDirParent() throws IOException {
+        // Note, that we can't reliably change the current working directory for the current java process
+        // therefore we use the current directory and make sure, we are at the correct place - in pmd-core
+        Path cwd = Paths.get(".").toRealPath();
+        assertThat(cwd.toString(), endsWith("pmd-core"));
+        String relativeSrcDir = IOUtil.normalizePath("src/test/resources/net/sourceforge/pmd/cli/src");
+        assertTrue(Files.isDirectory(cwd.resolve(relativeSrcDir)));
+
+        // use the parent directory
+        relativeSrcDir = relativeSrcDir + File.separator + "..";
+
+        startCapturingErrAndOut();
+        runPmd(StatusCode.VIOLATIONS_FOUND, "--no-cache", "--dir", relativeSrcDir, "--rulesets", DUMMY_RULESET_WITH_VIOLATIONS);
+
+        assertThat(outStreamCaptor.getLog(), containsString("\n" + relativeSrcDir + IOUtil.normalizePath("/src/somefile.dummy")));
+    }
+
+    @Test
+    public void testRelativizeWithRootRelativeSrcDir() throws IOException {
+        // Note, that we can't reliably change the current working directory for the current java process
+        // therefore we use the current directory and make sure, we are at the correct place - in pmd-core
+        Path cwd = Paths.get(".").toRealPath();
+        assertThat(cwd.toString(), endsWith("pmd-core"));
+        String relativeSrcDir = "src/test/resources/net/sourceforge/pmd/cli/src";
+        assertTrue(Files.isDirectory(cwd.resolve(relativeSrcDir)));
+
+        String root = cwd.getRoot().toString();
+
+        startCapturingErrAndOut();
+        runPmd(StatusCode.VIOLATIONS_FOUND, "--no-cache", "--dir", relativeSrcDir, "--rulesets", DUMMY_RULESET_WITH_VIOLATIONS, "--relativize-paths-with", root);
+
+        String absoluteSrcPath = cwd.resolve(relativeSrcDir).resolve("somefile.dummy").toString();
+        assertThat(outStreamCaptor.getLog(), containsString("\n" + absoluteSrcPath));
+    }
+
+    @Test
+    public void testRelativizeWith() {
+        startCapturingErrAndOut();
+        runPmd(StatusCode.VIOLATIONS_FOUND, "--no-cache", "--dir", srcDir, "--rulesets", DUMMY_RULESET_WITH_VIOLATIONS, "-z", srcDir.getParent());
+
+        assertThat(outStreamCaptor.getLog(), not(containsString(srcDir.resolve("someSource.dummy").toString())));
+        assertThat(outStreamCaptor.getLog(), startsWith(IOUtil.normalizePath("src/someSource.dummy")));
+    }
+
+    @Test
+    public void testRelativizeWithSymLink() throws IOException {
+        // srcDir = /tmp/junit123/src
+        // symlinkedSrcDir = /tmp/junit123/sources -> /tmp/junit123/src
+        Path symlinkedSrcDir = Files.createSymbolicLink(tempRoot().resolve("sources"), srcDir);
+        startCapturingErrAndOut();
+        runPmd(StatusCode.VIOLATIONS_FOUND, "--no-cache", "--dir", symlinkedSrcDir, "--rulesets", DUMMY_RULESET_WITH_VIOLATIONS, "-z", symlinkedSrcDir);
+
+        assertThat(outStreamCaptor.getLog(), not(containsString(srcDir.resolve("someSource.dummy").toString())));
+        assertThat(outStreamCaptor.getLog(), not(containsString(symlinkedSrcDir.resolve("someSource.dummy").toString())));
+        assertThat(outStreamCaptor.getLog(), startsWith("someSource.dummy"));
+    }
+
+    @Test
+    public void testRelativizeWithSymLinkParent() throws IOException {
+        // srcDir = /tmp/junit123/src
+        // symlinkedSrcDir = /tmp/junit-relativize-with-123 -> /tmp/junit123/src
+        Path tempPath = Files.createTempDirectory("junit-relativize-with-");
+        Files.delete(tempPath);
+        Path symlinkedSrcDir = Files.createSymbolicLink(tempPath, srcDir);
+        startCapturingErrAndOut();
+        // relativizing against parent of symlinkedSrcDir: /tmp
+        runPmd(StatusCode.VIOLATIONS_FOUND, "--no-cache", "--dir", symlinkedSrcDir, "--rulesets", DUMMY_RULESET_WITH_VIOLATIONS, "-z", symlinkedSrcDir.getParent());
+
+        assertThat(outStreamCaptor.getLog(), not(containsString(srcDir.resolve("someSource.dummy").toString())));
+        assertThat(outStreamCaptor.getLog(), not(containsString(symlinkedSrcDir.resolve("someSource.dummy").toString())));
+        // base path is symlinkedSrcDir without /tmp: e.g. junit-relativize-with-123
+        String basePath = symlinkedSrcDir.getParent().relativize(symlinkedSrcDir).toString();
+        assertThat(outStreamCaptor.getLog(), startsWith(basePath + File.separator + "someSource.dummy"));
+    }
+
+    @Test
+    public void testRelativizeWithMultiple() {
+        startCapturingErrAndOut();
+        runPmd(StatusCode.VIOLATIONS_FOUND, "--no-cache", "--dir", srcDir, "--rulesets", DUMMY_RULESET_WITH_VIOLATIONS, "-z", srcDir.getParent(), srcDir);
+
+        assertThat(outStreamCaptor.getLog(), not(containsString(srcDir.resolve("someSource.dummy").toString())));
+        assertThat(outStreamCaptor.getLog(), startsWith("someSource.dummy"));
+    }
+
+    @Test
+    public void testRelativizeWithFileIsError() {
+        startCapturingErrAndOut();
+        runPmd(StatusCode.ERROR, "--no-cache", "--dir", srcDir, "--rulesets", DUMMY_RULESET_WITH_VIOLATIONS, "-z", srcDir.resolve("someSource.dummy"));
+
+        assertThat(errStreamCaptor.getLog(), containsString(
+            "Expected a directory path for option --relativize-paths-with, found a file: "
+            + srcDir.resolve("someSource.dummy")));
+    }
+
+    @Test
+    public void testDeprecationWarningForShortNames() {
+        runPmdSuccessfully("--no-cache", "--dir", srcDir, "--rulesets", DUMMY_RULESET, "--short-names");
+
+        assertThat(loggingRule.getLog(), containsString("Some deprecated options were used on the command-line, including --short-names"));
+        assertThat(loggingRule.getLog(), containsString("Consider replacing it with --relativize-paths-with"));
     }
 
     @Test
@@ -315,6 +447,16 @@ public class CoreCliTest {
     private static void runPmd(StatusCode expectedExitCode, Object... args) {
         StatusCode actualExitCode = PMD.runPmd(argsToString(args));
         assertEquals("Exit code", expectedExitCode, actualExitCode);
+    }
+
+    public static class FooRule extends MockRule {
+
+        @Override
+        public void apply(List<? extends Node> nodes, RuleContext ctx) {
+            for (Node node : nodes) {
+                ctx.addViolation(node);
+            }
+        }
     }