+++ pmd-core/src/test/java/net/sourceforge/pmd/cli/PMDFilelistTest.java
@@ -6,7 +6,11 @@ package net.sourceforge.pmd.cli;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.endsWith;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.hasSize;
 
+import java.nio.file.Paths;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashSet;
@@ -18,18 +22,22 @@ import org.junit.Test;
 
 import net.sourceforge.pmd.PMD;
 import net.sourceforge.pmd.PMDConfiguration;
+import net.sourceforge.pmd.PmdAnalysis;
 import net.sourceforge.pmd.lang.DummyLanguageModule;
 import net.sourceforge.pmd.lang.Language;
+import net.sourceforge.pmd.lang.document.TextFile;
+import net.sourceforge.pmd.util.IOUtil;
 import net.sourceforge.pmd.util.datasource.DataSource;
 
 public class PMDFilelistTest {
+
+    private static final String RESOURCE_PREFIX = "src/test/resources/net/sourceforge/pmd/cli/";
+    private final Set<Language> languages = new HashSet<Language>(Arrays.asList(new DummyLanguageModule()));
+
     @Test
     public void testGetApplicableFiles() {
-        Set<Language> languages = new HashSet<>();
-        languages.add(new DummyLanguageModule());
-
         PMDConfiguration configuration = new PMDConfiguration();
-        configuration.setInputFilePath("src/test/resources/net/sourceforge/pmd/cli/filelist.txt");
+        configuration.setInputFilePath(RESOURCE_PREFIX + "filelist.txt");
 
         List<DataSource> applicableFiles = PMD.getApplicableFiles(configuration, languages);
         Assert.assertEquals(2, applicableFiles.size());
@@ -39,27 +47,76 @@ public class PMDFilelistTest {
 
     @Test
     public void testGetApplicableFilesMultipleLines() {
-        Set<Language> languages = new HashSet<>();
-        languages.add(new DummyLanguageModule());
-
         PMDConfiguration configuration = new PMDConfiguration();
-        configuration.setInputFilePath("src/test/resources/net/sourceforge/pmd/cli/filelist2.txt");
+        configuration.setInputFilePath(RESOURCE_PREFIX + "filelist2.txt");
 
         List<DataSource> applicableFiles = PMD.getApplicableFiles(configuration, languages);
-        Assert.assertEquals(3, applicableFiles.size());
+        Assert.assertEquals(2, applicableFiles.size());
         assertThat(applicableFiles.get(0).getNiceFileName(false, ""), endsWith("anotherfile.dummy"));
         assertThat(applicableFiles.get(1).getNiceFileName(false, ""), endsWith("somefile.dummy"));
-        assertThat(applicableFiles.get(2).getNiceFileName(false, ""), endsWith("somefile.dummy"));
+    }
+
+    @Test
+    public void testRelativizeWith() {
+        PMDConfiguration conf = new PMDConfiguration();
+        conf.setInputFilePath(Paths.get(RESOURCE_PREFIX + "filelist2.txt"));
+        conf.addRelativizeRoot(Paths.get("src/test/resources"));
+        try (PmdAnalysis pmd = PmdAnalysis.create(conf)) {
+            List<TextFile> files = pmd.files().getCollectedFiles();
+            assertThat(files, hasSize(2));
+            assertThat(files.get(0).getDisplayName(), equalTo(IOUtil.normalizePath("net/sourceforge/pmd/cli/src/anotherfile.dummy")));
+            assertThat(files.get(1).getDisplayName(), equalTo(IOUtil.normalizePath("net/sourceforge/pmd/cli/src/somefile.dummy")));
+        }
+    }
+
+    @Test
+    public void testRelativizeWithOtherDir() {
+        PMDConfiguration conf = new PMDConfiguration();
+        conf.setInputFilePath(Paths.get(RESOURCE_PREFIX + "filelist4.txt"));
+        conf.addRelativizeRoot(Paths.get(RESOURCE_PREFIX + "src"));
+        try (PmdAnalysis pmd = PmdAnalysis.create(conf)) {
+            List<TextFile> files = pmd.files().getCollectedFiles();
+            assertThat(files, hasSize(3));
+            assertThat(files.get(0).getDisplayName(), equalTo(".." + IOUtil.normalizePath("/otherSrc/somefile.dummy")));
+            assertThat(files.get(1).getDisplayName(), equalTo("anotherfile.dummy"));
+            assertThat(files.get(2).getDisplayName(), equalTo("somefile.dummy"));
+        }
+    }
+
+    @Test
+    public void testRelativizeWithSeveralDirs() {
+        PMDConfiguration conf = new PMDConfiguration();
+        conf.setInputFilePath(Paths.get(RESOURCE_PREFIX + "filelist4.txt"));
+        conf.addRelativizeRoot(Paths.get(RESOURCE_PREFIX + "src"));
+        conf.addRelativizeRoot(Paths.get(RESOURCE_PREFIX + "otherSrc"));
+        try (PmdAnalysis pmd = PmdAnalysis.create(conf)) {
+            List<TextFile> files = pmd.files().getCollectedFiles();
+            assertThat(files, hasSize(3));
+            assertThat(files.get(0).getDisplayName(), equalTo("somefile.dummy"));
+            assertThat(files.get(1).getDisplayName(), equalTo("anotherfile.dummy"));
+            assertThat(files.get(2).getDisplayName(), equalTo("somefile.dummy"));
+        }
+    }
+
+    @Test
+    public void testUseAbsolutePaths() {
+        PMDConfiguration conf = new PMDConfiguration();
+        conf.setInputFilePath(Paths.get(RESOURCE_PREFIX + "filelist4.txt"));
+        conf.addRelativizeRoot(Paths.get(RESOURCE_PREFIX).toAbsolutePath().getRoot());
+        try (PmdAnalysis pmd = PmdAnalysis.create(conf)) {
+            List<TextFile> files = pmd.files().getCollectedFiles();
+            assertThat(files, hasSize(3));
+            assertThat(files.get(0).getDisplayName(), equalTo(Paths.get(RESOURCE_PREFIX, "otherSrc", "somefile.dummy").toAbsolutePath().toString()));
+            assertThat(files.get(1).getDisplayName(), equalTo(Paths.get(RESOURCE_PREFIX, "src", "anotherfile.dummy").toAbsolutePath().toString()));
+            assertThat(files.get(2).getDisplayName(), equalTo(Paths.get(RESOURCE_PREFIX, "src", "somefile.dummy").toAbsolutePath().toString()));
+        }
     }
 
     @Test
     public void testGetApplicatbleFilesWithIgnores() {
-        Set<Language> languages = new HashSet<>();
-        languages.add(new DummyLanguageModule());
-
         PMDConfiguration configuration = new PMDConfiguration();
-        configuration.setInputFilePath("src/test/resources/net/sourceforge/pmd/cli/filelist3.txt");
-        configuration.setIgnoreFilePath("src/test/resources/net/sourceforge/pmd/cli/ignorelist.txt");
+        configuration.setInputFilePath(RESOURCE_PREFIX + "filelist3.txt");
+        configuration.setIgnoreFilePath(RESOURCE_PREFIX + "ignorelist.txt");
 
         List<DataSource> applicableFiles = PMD.getApplicableFiles(configuration, languages);
         Assert.assertEquals(2, applicableFiles.size());
@@ -69,12 +126,9 @@ public class PMDFilelistTest {
 
     @Test
     public void testGetApplicatbleFilesWithDirAndIgnores() {
-        Set<Language> languages = new HashSet<>();
-        languages.add(new DummyLanguageModule());
-
         PMDConfiguration configuration = new PMDConfiguration();
-        configuration.setInputPaths("src/test/resources/net/sourceforge/pmd/cli/src");
-        configuration.setIgnoreFilePath("src/test/resources/net/sourceforge/pmd/cli/ignorelist.txt");
+        configuration.setInputPaths(RESOURCE_PREFIX + "src");
+        configuration.setIgnoreFilePath(RESOURCE_PREFIX + "ignorelist.txt");
 
         List<DataSource> applicableFiles = PMD.getApplicableFiles(configuration, languages);
         Assert.assertEquals(4, applicableFiles.size());